requirements.txt
aiohttp==3.7.4.post0
allennlp==2.7.0
antlr4-python3-runtime==4.9.3
argon2-cffi==21.1.0
astroid==2.8.4
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
backports.csv==1.0.7
base58==2.1.0
beautifulsoup4==4.10.0
bleach==4.1.0
blessings==1.7
blis==0.7.4
boto3==1.18.62
botocore==1.21.62
bz2file==0.98
cachetools==4.2.4
catalogue==2.0.6
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
checklist==0.0.11
cheroot==8.5.2
CherryPy==18.6.1
click==8.0.3
colorama==0.4.6
commonmark==0.9.1
configparser==5.0.2
cryptography==35.0.0
cymem==2.0.5
datasets==1.13.2
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
entrypoints==0.3
fairscale==0.4.0
FastNLP==0.5.0
feedparser==6.0.8
filelock==3.0.12
fonttools==4.25.0
fsspec==2021.10.0
future==0.18.2
gensim==4.1.2
gitdb==4.0.7
GitPython==3.1.24
google-api-core==2.1.1
google-auth==2.3.0
google-cloud-core==2.1.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
gpustat==0.6.0
h5py==3.4.0
huggingface-hub==0.0.19
idna==3.3
iniconfig==1.1.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
iso-639==0.4.5
isort==5.9.3
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.functools==3.3.0
jaraco.text==3.5.1
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jsonnet==0.17.0
jsonschema==4.1.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lazy-object-proxy==1.6.0
lmdb==1.2.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-service==2.4.0
more-itertools==8.10.0
multidict==5.2.0
multiprocess==0.70.14
munch==2.5.0
munkres==1.1.4
murmurhash==1.0.5
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.5
notebook==6.4.4
nvidia-ml-py3==7.352.0
omegaconf==2.2.3
overrides==3.1.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pathos==0.3.0
pathtools==0.1.2
pathy==0.6.0
patternfork-nosql==3.6
pdfminer.six==20211012
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.0
pluggy==1.0.0
portend==3.0.0
pox==0.3.2
ppft==1.7.6.6
preshed==3.0.5
prettytable==3.5.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.18.1
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
pylint==2.11.1
pyrsistent==0.18.0
pytest==6.2.5
python-docx==0.8.11
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.10.8
requests==2.26.0
rich==11.2.0
rsa==4.7.2
s3transfer==0.5.0
sacremoses==0.0.46
scikit-learn==1.0
scipy==1.7.1
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.4.3
sgmllib3k==1.0.0
shortuuid==1.0.1
sip==4.19.13
smart-open==5.2.1
smmap==4.0.0
soupsieve==2.2.1
spacy==3.1.3
spacy-legacy==3.0.8
sqlitedict==1.7.0
srsly==2.4.1
subprocess32==3.5.4
tempora==4.1.2
tensorboardX==2.4
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thinc==8.0.10
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3
traitlets==5.1.0
transformers==4.5.1
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wandb==0.12.4
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.13.2
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
xxhash==2.0.2
yapf==0.31.0
yarl==1.7.0
yaspin==2.1.0
zc.lockfile==2.0
